[TS-335] Refactor: create habit page refactoring
