+++ src/components/Habit/CreateHabit/HabitTip/HabitTip.style.ts
