+++ src/components/Habit/CreateHabit/habitGenerateConditions.tsx
