+++ src/components/Habit/Profile/Profile.tsx
@@ -7,6 +7,8 @@ import StarImage from "@/assets/image/img-card-detail-star-button.png";
 import { useAppDispatch, useAppSelector } from "@/api/hooks";
 import { getUserInfo } from "@/api/user/userThunk";
 
+import { PATH } from "@/constants/path";
+
 import {
 	layoutStyle,
 	profileBoxStyle,
@@ -27,14 +29,14 @@ function Profile() {
 
 	return (
 		<div css={layoutStyle}>
-			<div css={profileBoxStyle} onClick={() => navigate("/mypage")}>
+			<div css={profileBoxStyle} onClick={() => navigate(PATH.MY)}>
 				<img src={StarImage} alt="user 프로필 이미지" />
 				<div>
 					<p css={identityTextStyle}>{userData.identity}</p>
 					<p css={nicknameTextStyle}>{userData.nickname}</p>
 				</div>
 			</div>
-			<ProfileButtonIcon onClick={() => navigate("/habit-generate")} />
+			<ProfileButtonIcon onClick={() => navigate(PATH.CREATE_HABIT)} />
 		</div>
 	);
 }
