+++ src/constants/path.ts
@@ -1,3 +1,5 @@
 export const PATH = {
 	CHATTING: "/chatting",
+	CREATE_HABIT: "/habit-new",
+	MY: "/mypage",
 } as const;
